Extract `container` and `service` modules from `main.rs` in `oct-ctl`
It is the initial step for #250
